fix: split campaign provenance contracts
